requirements.txt
Django>=2.0,<2.1
djangorestframework
